@@ -1,17 +1,24 @@
 import { SearchPatternType, SearchMode } from '@sourcegraph/search'
-import { aggregateStreamingSearch, LATEST_VERSION } from '@sourcegraph/shared/src/search/stream'
+import {
+    aggregateStreamingSearch,
+    LATEST_VERSION,
+    type AggregateStreamingSearchResults,
+} from '@sourcegraph/shared/src/search/stream'
 import type { StreamSearchOptions } from '@sourcegraph/shared/src/search/stream'
-import { BehaviorSubject, merge } from 'rxjs'
-import { last, share, throttleTime } from 'rxjs/operators/index'
+import { BehaviorSubject, merge, Observable, of } from 'rxjs'
+import { last, share, shareReplay, throttleTime } from 'rxjs/operators/index'
 import type { PageLoad } from './$types'
-import { browser } from '$app/environment'
+import { navigating } from '$app/stores'
+import { get } from 'svelte/store'
 
-export const load: PageLoad = ({ url }) => {
+const cache: Record<string, Observable<AggregateStreamingSearchResults | undefined>> = {}
+
+export const load: PageLoad = ({ url, depends }) => {
     const query = url.searchParams.get('q')
 
-    console.log('browser', browser)
-
     if (query) {
+        depends(`query:${query}`)
+
         const options: StreamSearchOptions = {
             version: LATEST_VERSION,
             patternType: SearchPatternType.literal,
@@ -20,17 +27,37 @@ export const load: PageLoad = ({ url }) => {
             featureOverrides: [],
             searchMode: SearchMode.Precise,
             chunkMatches: true,
-            sourcegraphURL: 'https://sourcegraph.test:3443',
         }
 
-        const querySource = new BehaviorSubject<string>(query)
-        const searchStream = aggregateStreamingSearch(querySource, options).pipe(share())
-        const resultStream = merge(searchStream.pipe(throttleTime(500)), searchStream.pipe(last()))
+        const key = createCacheKey(options, query)
+        let searchStream = cache[key]
+        if (get(navigating)?.type !== 'popstate' || !searchStream) {
+            const querySource = new BehaviorSubject<string>(query)
+            searchStream = cache[key] = merge(of(undefined), aggregateStreamingSearch(querySource, options)).pipe(
+                shareReplay(1)
+            )
+            // Primes the stream
+            searchStream.subscribe()
+        }
+        const resultStream = searchStream //merge(searchStream.pipe(throttleTime(500)), searchStream.pipe(last()))
 
         return {
+            query,
             stream: resultStream,
         }
     } else {
-        return { stream: null }
+        return { query, stream: null }
     }
+}
+
+function createCacheKey(options: StreamSearchOptions, query: string): string {
+    return [
+        options.version,
+        options.patternType,
+        options.caseSensitive,
+        options.caseSensitive,
+        options.searchMode,
+        options.chunkMatches,
+        query,
+    ].join('--')
 }
