@@ -6,8 +6,8 @@
     export let data: PageData
 </script>
 
-{#if data.stream}
-    <SearchResults stream={data.stream} />
+{#if data.query && data.stream}
+    <SearchResults stream={data.stream} query={data.query} />
 {:else}
     <SearchHome />
 {/if}
