@@ -6,11 +6,10 @@ import * as GQL from '@sourcegraph/shared/src/schema'
 
 const getHeaders = (): { [header: string]: string } => {
     const headers: { [header: string]: string } = {
-        //...window?.context?.xhrHeaders,
+        ...window?.context?.xhrHeaders,
         Accept: 'application/json',
         'Content-Type': 'application/json',
     }
-    /*
     const parameters = new URLSearchParams(window.location.search)
     const trace = parameters.get('trace')
     if (trace) {
@@ -20,7 +19,6 @@ const getHeaders = (): { [header: string]: string } => {
     if (feat.length) {
         headers['X-Sourcegraph-Override-Feature'] = feat.join(',')
     }
-    */
     return headers
 }
 
@@ -41,7 +39,6 @@ export const requestGraphQL = <TResult, TVariables = object>(
         request,
         variables,
         headers: getHeaders(),
-        baseUrl: 'https://sourcegraph.test:3443',
     })
 
 /**
